requirements.txt
Django==4.2.1
# django-cors-middleware===1.5.0
djangorestframework==3.14.0
gunicorn==20.1.0
mysqlclient==2.1.1
pytest-cov==4.0.0
pytest-django==4.5.2
whitenoise==6.4.0